requirements.txt
alembic==1.7.6
Babel==2.9.1
blinker==1.4
click==8.0.4
dnspython==2.2.1
email-validator==1.1.3
Flask==2.0.3
Flask-BabelEx==0.9.4
Flask-CKEditor==0.4.6
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Principal==0.4.0
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
Mako==1.2.0
MarkupSafe==2.1.0
passlib==1.7.4
python-slugify==6.1.1
pytz==2021.3
speaklater==1.3
SQLAlchemy==1.4.32
text-unidecode==1.3
Werkzeug==2.0.3
WTForms==3.0.1
